move `useLgTemplate` under hooks folder
